requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.2
